+++ src/main/java/vendingmachine/model/domain/Coin.java
@@ -1,4 +1,4 @@
-package vendingmachine;
+package vendingmachine.model.domain;
 
 import java.util.ArrayList;
 import java.util.List;
