+++ src/test/java/vendingmachine/model/domain/VendingMachineTest.java
@@ -4,8 +4,6 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import vendingmachine.Coin;
-
 import java.util.Map;
 
 import static org.assertj.core.api.Assertions.assertThat;
